implemented open and make objects visible
