+++ igibson/object_states/robot_related_states.py
@@ -1,10 +1,14 @@
 import numpy as np
 import pyquaternion # import for custom code 
+from scipy.spatial import ConvexHull
+from shapely.geometry import Polygon
 
 from igibson.object_states.object_state_base import BooleanState, CachingEnabledObjectState
 from igibson.object_states.pose import Pose
 from igibson.object_states.room_states import InsideRoomTypes
 from igibson.utils.constants import MAX_INSTANCE_COUNT
+
+import igibson.render_utils as render_utils
 
 _IN_REACH_DISTANCE_THRESHOLD = 2.0
 
@@ -22,6 +26,68 @@ def _get_robot(simulator):
 
     return valid_robots[0]
 
+def set_camera(s,
+               robot,
+               field_of_view=120,
+               forward_downward_dir=None,
+               up_dir=None,
+               camera_pos_offset=None,
+               apply_q_rotation=True
+              ):
+    # Set default values
+    if forward_downward_dir is None:
+        forward_downward_dir = np.array([1, 0, -0.25])
+    if up_dir is None:
+        up_dir = np.array([0, 0, 1])
+    if camera_pos_offset is None:
+        camera_pos_offset = np.array([0.1, 0.1, 0.7])
+
+    robot_pos, q = get_robot_pos_and_q_rotation(robot)
+
+    if apply_q_rotation:
+        # Apply rotations to see all directions in the frame of the robot
+        camera_pos = robot_pos + q.rotate(camera_pos_offset) # Slightly above the robot's center
+        forward_downward_direction = q.rotate(forward_downward_dir) # Forward wrt the robot's frame of reference 
+        up_direction = up_dir # up is up no matter the frame of reference, unless the robot is doing something weird
+    else:
+        camera_pos = robot_pos + camera_pos_offset
+        forward_downward_direction = forward_downward_dir
+        up_direction = up_dir
+
+    # Set the camera in the renderer
+    s.renderer.set_camera(camera_pos, camera_pos + forward_downward_direction, up_direction)
+    s.renderer.set_fov(field_of_view)
+
+def get_robot_pos_and_q_rotation(robot):
+    robot_pos, robot_orientation = robot.get_position_orientation()
+    
+    # Convert quaternion to rotation matrix - takes w,x,y,z in input, but robot orientation is given as x,y,z,w !!!
+    q = pyquaternion.Quaternion(x=robot_orientation[0], 
+                                y=robot_orientation[1], 
+                                z=robot_orientation[2], 
+                                w=robot_orientation[3])
+    return robot_pos, q
+
+
+def compute_projected_area(uv_coords):
+    """
+    uv_coords: (8, 2) array of projected 2D points (u,v) in image coordinates
+    returns: area in pixels² of the convex hull of the projection
+    """
+    uv_coords = np.asarray(uv_coords)
+
+    if uv_coords.shape[0] < 3:
+        return 0  # Not enough points to define an area
+
+    try:
+        hull = ConvexHull(uv_coords)
+        hull_pts = uv_coords[hull.vertices]
+        polygon = Polygon(hull_pts)
+        return polygon.area
+    except Exception as e:
+        # Never encontered an error so far
+        print(f"Warning: convex hull failed with error: {e}")
+        return 1 # return minimum number of pixels possible that is not illegal as denominator
 
 class InReachOfRobot(CachingEnabledObjectState, BooleanState):
     @staticmethod
@@ -116,10 +182,7 @@ class InFOVOfRobot(CachingEnabledObjectState, BooleanState):
             return False
 
         body_ids = set(self.obj.get_body_ids())
-        #print(robot.states)
-        #print(ObjectsInFOVOfRobot)
-        #print(robot.states[ObjectsInFOVOfRobot])
-        return not body_ids.isdisjoint(robot.states[ObjectsInFOVOfRobot].get_value()) # this one is crushing badly
+        return not body_ids.isdisjoint(robot.states[ObjectsInFOVOfRobot].get_value()) 
 
     def _set_value(self, new_value):
         raise NotImplementedError("InFOVOfRobot state currently does not support setting.")
@@ -131,6 +194,64 @@ class InFOVOfRobot(CachingEnabledObjectState, BooleanState):
     def load(self, data):
         pass
 
+
+
+class IsVisible(CachingEnabledObjectState, BooleanState):
+    @staticmethod
+    def get_optional_dependencies():
+        return CachingEnabledObjectState.get_optional_dependencies() + [VisiblePixelCountFromRobot] # no idea why this is needed!
+
+    def _compute_value(self, env, relative_threshold=0.08, absolute_threshold=None, use_absolute_threshold=False):
+        """
+        Note that this function has a mandatory argument env!
+        """
+        robot = _get_robot(self.simulator)
+        if not robot:
+            return False
+
+        body_ids = set(self.obj.get_body_ids())
+
+        # Hopefully self.obj.states is the right thing to access - double check!
+        total_visible_pixels = self.obj.states[VisiblePixelCountFromRobot].get_value()
+        
+        if use_absolute_threshold:
+            if absolute_threshold is None:
+                absolute_threshold = MIN_PIXEL_VISIBLE
+            # Return True if the amount of visible pixels from all body parts of the object are more than absolute_threshold
+            # Else return False
+            return total_visible_pixels > absolute_threshold
+        else:
+            # In the relative case, take the ratio between the total visible pixels and the area of the convex hull
+            # formed by the vertices of the object bounding box. As this is not a tight box, ratios of ~8% are usually 
+            # enough to consider an object visible
+            
+            bbox_vertices_uv = render_utils.render_frame_with_trg_obj(env, 
+                                                                      self.obj, 
+                                                                      show=False, 
+                                                                      save=False, 
+                                                                      add_bbox=True, 
+                                                                      return_uv_coord=True
+                                                                     )
+            
+            projected_bbox_pixels = compute_projected_area(bbox_vertices_uv)
+            return (total_visible_pixels / (projected_bbox_pixels + 1)) >= relative_threshold 
+
+    def _set_value(self, new_value):
+        raise NotImplementedError("IsVisible state currently does not support setting.")
+
+    def _get_value(self, *args, **kwargs):
+        # If we don't have a value cached, compute it now.
+        if self.value is None:
+            self.value = self._compute_value(*args, **kwargs)
+
+        return self.value
+        
+    # Nothing to do here.
+    def _dump(self):
+        pass
+
+    def load(self, data):
+        pass
 
 class ObjectsInFOVOfRobot(CachingEnabledObjectState):
     def __init__(self, *args, **kwargs):
@@ -189,45 +310,29 @@ class ObjectsInFOVOfRobot(CachingEnabledObjectState):
             frames.append(item)
         return frames
 
+class VisiblePixelCountFromRobot(ObjectsInFOVOfRobot):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
 
-def set_camera(s,
-               robot,
-               field_of_view=120,
-               forward_downward_dir=None,
-               up_dir=None,
-               camera_pos_offset=None,
-               apply_q_rotation=True
-              ):
-    # Set default values
-    if forward_downward_dir is None:
-        forward_downward_dir = np.array([1, 0, -0.25])
-    if up_dir is None:
-        up_dir = np.array([0, 0, 1])
-    if camera_pos_offset is None:
-        camera_pos_offset = np.array([0.1, 0.1, 0.7])
+    def _compute_value(self):
+        # Render instance segmentation
+        seg = self.render_single_robot_camera(modes="ins_seg")[0][:, :, 0]
+        seg = np.round(seg * MAX_INSTANCE_COUNT).astype(int)
+        body_ids = self.simulator.renderer.get_pb_ids_for_instance_ids(seg)
 
-    robot_pos, q = get_robot_pos_and_q_rotation(robot)
+        flat_ids = body_ids.flatten()
 
-    if apply_q_rotation:
-        # Apply rotations to see all directions in the frame of the robot
-        camera_pos = robot_pos + q.rotate(camera_pos_offset) # Slightly above the robot's center
-        forward_downward_direction = q.rotate(forward_downward_dir) # Forward wrt the robot's frame of reference 
-        up_direction = up_dir # up is up no matter the frame of reference, unless the robot is doing something weird
-    else:
-        camera_pos = robot_pos + camera_pos_offset
-        forward_downward_direction = forward_downward_dir
-        up_direction = up_dir
+        # Count pixels per body ID
+        unique_ids, counts = np.unique(flat_ids, return_counts=True)
+        pixel_counts = dict(zip(unique_ids, counts))
 
-    # Set the camera in the renderer
-    s.renderer.set_camera(camera_pos, camera_pos + forward_downward_direction, up_direction)
-    s.renderer.set_fov(field_of_view)
+        total_visible_pixels = 0
+        obj_body_ids = set(self.obj.get_body_ids())
 
-def get_robot_pos_and_q_rotation(robot):
-    robot_pos, robot_orientation = robot.get_position_orientation()
-    
-    # Convert quaternion to rotation matrix - takes w,x,y,z in input, but robot orientation is given as x,y,z,w !!!
-    q = pyquaternion.Quaternion(x=robot_orientation[0], 
-                                y=robot_orientation[1], 
-                                z=robot_orientation[2], 
-                                w=robot_orientation[3])
-    return robot_pos, q
+        for body_id in obj_body_ids:
+            total_visible_pixels += pixel_counts.get(body_id, 0)
+
+        return total_visible_pixels
+
+
+
